\name{neighbCount}
\alias{neighbCount3d}
\alias{neighbCount2d}
\title{
Neighbours count in radius x.
}
\description{
neighbCount
}
\usage{
neighbCount2d(x, y, radius = 0.05)
neighbCount3d(x, y, z, radius = 0.05)
}
\arguments{
  \item{x}{vector of x coordinate}
  \item{y}{vector of y coordinate}
  \item{z}{vector of z coordinate}
  \item{radius}{radius in meters}
}
\details{
Calculates neighbours count in a given radius.
}
