\name{neighbCount}
\alias{neighbCount}
\title{
Neighbours count in radius x.
}
\description{
neighbCount
}
\usage{
neighbCount(x, y, radius = 0.05)
}
\arguments{
  \item{x}{vector of x coordinate}
  \item{y}{vector of y coordinate}
  \item{radius}{radius in meters}
}
\details{
Calculates neighbours count in a given radius.
}
